sdfsdfsdf

+++ projects/CO-DETR/codetr/co_dino_head.py
@@ -639,7 +639,7 @@ class CoDINOHead(DINOHead):
         scores[pos_inds] = bbox_overlaps(
             pos_decode_bbox_pred.detach(),
             pos_decode_bbox_targets,
-            is_aligned=True)
+            is_aligned=True).to(scores.dtype)
         loss_cls = self.loss_cls(
             cls_scores, (labels, scores),
             weight=label_weights,
